+++ src/main/java/io/github/thelordman/costrength/utilities/Data.java
@@ -20,9 +20,9 @@ public class Data implements Serializable {
     public static final HashMap<UUID, FastBoard> scoreboard = new HashMap<>();
     public static final HashMap<Player, Byte> combatTag = new HashMap<>();
 
-    public final HashMap<UUID, Float> balance = new HashMap<>();
-    public final HashMap<UUID, Float> bounty = new HashMap<>();
-    public final HashMap<UUID, Float> xp = new HashMap<>();
+    public final HashMap<UUID, Double> balance = new HashMap<>();
+    public final HashMap<UUID, Double> bounty = new HashMap<>();
+    public final HashMap<UUID, Double> xp = new HashMap<>();
     public final HashMap<UUID, Integer> level = new HashMap<>();
     public final HashMap<UUID, Integer> killstreak = new HashMap<>();
 }
